requirements.txt
streamlit==1.31.0
langchain==0.1.0
openai==1.12.0
pinecone-client==3.0.0
python-dotenv==1.0.0
tiktoken==0.5.2